requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
diff-match-patch==20241021
Django==5.1.5
django-import-export==4.3.3
et_xmlfile==2.0.0
frozenlist==1.5.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.1.0
kombu==5.4.2
multidict==6.1.0
numpy==2.2.1
openpyxl==3.1.5
pandas==2.2.3
prompt_toolkit==3.0.48
propcache==0.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
schedule==1.2.2
six==1.17.0
SkPy==0.11
soupsieve==2.6
sqlparse==0.5.3
tablib==3.7.0
telepot==12.7
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
